Two sections were deleted.
First, the section importing Kazuza database codon usage tables was deleted and moved to the Statistical_Analysis_of_Codon_Usages.R file. This was done becasue the resulting imported data frames were not being analyzed in this file. (i.e. This file only builds a codon usage table for a given species of interest. The Statistical_Analysis_of_Codon_Usages.R performs comparisons between different codon usages).

Second, a portion from the section examining coding sequences lengths was deleted because there is no evidence to exculde a coding sequence simply because it is an "outlier" from an average sequence length. This examination of sequence lengths is crucial when applying the MILC statistical measure and can be considered in the statistical analysis portion of the next script.