Removed the section on Intra-samples Karlin B Plot and associated statistical work. This was done for three main reasons.
First, the entire genome of Nicotiana benthamaina and all of the gene sequences has not been fully sequenced. Thus, any characteristic crescent moon pattern in MILC, B, MCB, and ENC Prime distance plots may not be visible.

Second, the relative "complexity" of this figure is not necessary to convey the similariries and differences between samples. By computing and comparing the MILC distance in the same species - as a new and old codon usage - the data points of the codon usage bias measure would not be expected to differ. Thus, it does not warrant a intra-sample comparison & does not make for an informative figure.

Lastly, similar information to this figure can be conveyed in a more easy-to-understand manner. Some examples include: reporting the B measure between old and new codon usages as a measure of codon usage bias OR reporting the RSCU values to understand codon usage preference differences.